# internal/docs/swagger.yaml
definitions:
  models.BlockReward:
    properties:
      reward:
        type: integer
      status:
        type: boolean
    type: object
  models.Error:
    properties:
      error:
        type: string
    type: object
  models.SyncDuties:
    properties:
      validators:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
paths:
  /blockreward/{slot}:
    get:
      consumes:
      - application/json
      description: Get the reward for a specific slot
      parameters:
      - description: Slot Number
        in: path
        name: slot
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BlockReward'
        "400":
          description: slot is in the future / invalid request params
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: the slot does not exist / was missed
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get slot reward
      tags:
      - rewards
  /syncduties/{slot}:
    get:
      consumes:
      - application/json
      description: Get the pubkeys of the validators in the sync committee for a specific
        slot
      parameters:
      - description: Slot Number
        in: path
        name: slot
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SyncDuties'
        "400":
          description: slot is in the future / invalid request params
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: the slot does not exist / was missed
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get sync duties for given slot
      tags:
      - syncduties
swagger: "2.0"